+++ backend/src/main/java/edonymyeon/backend/notification/application/NotificationSender.java
@@ -1,5 +1,7 @@
 package edonymyeon.backend.notification.application;
 
+import edonymyeon.backend.notification.application.dto.Receiver;
+
 public interface NotificationSender {
     void sendNotification(Receiver receiver, String title);
 }
